+++ src/datastore/index.js
@@ -100,7 +100,6 @@ defaultsPrototype.eagerEject = false;
 defaultsPrototype.eagerInject = false;
 defaultsPrototype.endpoint = '';
 defaultsPrototype.error = console ? (a, b, c) => console[typeof console.error === 'function' ? 'error' : 'log'](a, b, c) : false;
-defaultsPrototype.instanceEvents = !!DSUtils.w;
 defaultsPrototype.fallbackAdapters = ['http'];
 defaultsPrototype.findBelongsTo = true;
 defaultsPrototype.findHasOne = true;
@@ -109,7 +108,7 @@ defaultsPrototype.findInverseLinks = true;
 defaultsPrototype.findStrictCache = true;
 defaultsPrototype.idAttribute = 'id';
 defaultsPrototype.ignoredChanges = [/\$/];
-defaultsPrototype.ignoreMissing = false;
+defaultsPrototype.instanceEvents = !!DSUtils.w;
 defaultsPrototype.keepChangeHistory = false;
 defaultsPrototype.linkRelations = true;
 defaultsPrototype.log = console ? (a, b, c, d, e) => console[typeof console.info === 'function' ? 'info' : 'log'](a, b, c, d, e) : false;
@@ -129,8 +128,8 @@ defaultsPrototype.onConflict = 'merge';
 defaultsPrototype.reapAction = !!DSUtils.w ? 'inject' : 'none';
 defaultsPrototype.reapInterval = !!DSUtils.w ? 30000 : false;
 defaultsPrototype.relationsEnumerable = false;
-defaultsPrototype.returnMeta = false;
 defaultsPrototype.resetHistoryOnInject = true;
+defaultsPrototype.returnMeta = false;
 defaultsPrototype.strategy = 'single';
 defaultsPrototype.upsert = !!DSUtils.w;
 defaultsPrototype.useClass = true;
@@ -336,6 +335,8 @@ class DS {
         });
       };
     }
+
+    DSUtils.Events(_this);
   }
 
   getAdapterName(options) {